requirements.txt
3==0.0.0
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
environs==10.3.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
marshmallow==3.20.2
multidict==6.0.5
packaging==23.2
peewee==3.17.0
pydantic==2.5.3
pydantic_core==2.14.6
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
